@@ -1,6 +1,6 @@
 module github.com/Shopify/go-storage
 
-go 1.17
+go 1.20
 
 require (
 	cloud.google.com/go/storage v1.23.0
