@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 )
 
 func Read(ctx context.Context, fs FS, path string, options *ReaderOptions) ([]byte, error) {
@@ -16,7 +15,7 @@ func Read(ctx context.Context, fs FS, path string, options *ReaderOptions) ([]by
 	}
 
 	var data []byte
-	if data, err = ioutil.ReadAll(file); err != nil {
+	if data, err = io.ReadAll(file); err != nil {
 		_ = file.Close() // Best effort at cleaning up
 
 		return nil, fmt.Errorf("unable to read %s: %w", path, err)
